# Add a C implementation for `TensorFromScalar`