Closes #1123

While inspecting the generated C code, I noticed that there might be an unnecessary call to `PyArray_ScalarAsCtype` introduced by `ScalarType.c_extract`.