@mattearllongshot 2022-08-19
This is much faster, thank you:

Before  (9f176da71d41635e8854fd601fd6a68102b0c6e5)
```
Ops
---
<% time> <sum %> <apply time> <time per call> <type> <#call> <#apply> <Op name>
  19.4%    19.4%       0.331s       1.00e-05s     C     33000        3   CGemv{inplace}
  17.7%    37.0%       0.302s       1.37e-05s     Py    22000        2   BroadcastTo
   9.7%    46.8%       0.166s       1.51e-05s     C     11000        1   Elemwise{...}
   7.2%    54.0%       0.123s       1.12e-05s     C     11000        1   Elemwise{...}[(0, 3)]
   6.9%    60.9%       0.118s       2.15e-06s     C     55000        5   IncSubtensor{InplaceInc;int64}
   5.5%    66.4%       0.094s       2.14e-06s     Py    44000        4   TensorFromScalar
   2.4%    68.8%       0.041s       3.69e-06s     C     11000        1   Softmax{axis=-1}
   1.8%    70.6%       0.031s       2.00e-07s     C     154000       14   Switch
   1.8%    72.4%       0.031s       5.56e-07s     C     55000        5   InplaceDimShuffle{x,0}
   1.7%    74.1%       0.029s       1.34e-06s     C     22000        2   Alloc
   1.7%    75.8%       0.029s       1.89e-07s     C     154000       14   EQ
   1.3%    77.2%       0.023s       6.84e-07s     C     33000        3   InplaceDimShuffle{0,x}
   1.3%    78.5%       0.022s       1.02e-06s     C     22000        2   Elemwise{...}
   1.3%    79.8%       0.022s       2.01e-06s     C     11000        1   SoftmaxGrad{axis=-1}
   1.3%    81.0%       0.022s       1.82e-07s     C     121000       11   maximum
   1.3%    82.3%       0.022s       4.98e-07s     C     44000        4   InplaceDimShuffle{x}
   1.3%    83.6%       0.022s       2.18e-07s     C     99000        9   Shape_i{0}
   1.2%    84.8%       0.021s       2.75e-07s     C     77000        7   Sum{acc_dtype=float64}
   0.9%    85.7%       0.015s       1.39e-07s     C     110000       10   Abs
   0.9%    86.6%       0.015s       4.52e-07s     C     33000        3   AllocEmpty{dtype='float64'}
   ... (remaining 31 Ops account for  13.40%(0.23s) of the runtime)
```

After  (0a16c225ef8f2de1205bc1494a4ecb50108ce4b6)
```
Ops
---
<% time> <sum %> <apply time> <time per call> <type> <#call> <#apply> <Op name>
  19.8%    19.8%       0.342s       1.55e-05s     Py    22000        2   BroadcastTo
  19.6%    39.4%       0.339s       1.03e-05s     C     33000        3   CGemv{inplace}
   9.9%    49.3%       0.170s       1.55e-05s     C     11000        1   Elemwise{...}
   7.6%    56.8%       0.131s       2.38e-06s     C     55000        5   IncSubtensor{InplaceInc;int64}
   7.2%    64.1%       0.125s       1.13e-05s     C     11000        1   Elemwise{...}[(0, 3)]
   2.4%    66.5%       0.041s       3.77e-06s     C     11000        1   Softmax{axis=-1}
   1.9%    68.4%       0.033s       1.52e-06s     C     22000        2   Alloc
   1.8%    70.2%       0.031s       2.03e-07s     C     154000       14   Switch
   1.8%    72.0%       0.031s       5.66e-07s     C     55000        5   InplaceDimShuffle{x,0}
   1.8%    73.8%       0.031s       2.01e-07s     C     154000       14   EQ
   1.4%    75.2%       0.025s       7.58e-07s     C     33000        3   InplaceDimShuffle{0,x}
   1.4%    76.6%       0.024s       2.02e-07s     C     121000       11   maximum
   1.4%    78.0%       0.024s       5.45e-07s     C     44000        4   InplaceDimShuffle{x}
   1.4%    79.4%       0.023s       3.05e-07s     C     77000        7   Sum{acc_dtype=float64}
   1.3%    80.7%       0.023s       2.11e-06s     C     11000        1   SoftmaxGrad{axis=-1}
   1.3%    82.0%       0.022s       1.01e-06s     C     22000        2   Elemwise{...}
   1.3%    83.3%       0.022s       2.20e-07s     C     99000        9   Shape_i{0}
   1.0%    84.3%       0.018s       5.47e-07s     C     33000        3   AllocEmpty{dtype='float64'}
   0.9%    85.2%       0.016s       3.61e-07s     C     44000        4   TensorFromScalar
   0.9%    86.2%       0.016s       1.42e-07s     C     110000       10   Abs
   ... (remaining 31 Ops account for  13.85%(0.24s) of the runtime)
```